Created a new cluster-scoped operator project
This was created by using the command:

```
operator-sdk init \
    --domain cyberark.com \
    --repo github.com/cyberark/conjur-config-operator
    --skip-go-version-check
```

The reference used was the `Create a new project` section of the
Operator-SDK Golang tutorial
(https://sdk.operatorframework.io/docs/building-operators/golang/tutorial/#create-a-new-project).

The operator is left as a "cluster-scoped" operator (see
https://sdk.operatorframework.io/docs/building-operators/golang/operator-scope/),
which is the default for the `operator-sdk init ...` command, so that the
operator can watch Conjur Config custom resources in all Namespaces.